+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 This file is used to list changes made in each version of grafana.
 
+## Unreleased
+
 ## 10.0.6 - *2022-10-12*
 
 - Correct `config_ldap_group_mapping` documentation file name