+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This file is used to list changes made in each version of grafana.
 
-## Unreleased
+## 10.2.0 - *2022-12-01*
 
 - Add `enabled` property to `config_unified_alerting` resource [grafana docs](https://grafana.com/docs/grafana/latest/alerting/migrating-alerts/roll-back/)
 - Add `force_migration` property to `config` resource to allow unified alerting to be rolled back